backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# 数据库
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
neo4j==5.19.0
tortoise-orm==0.22.0

# Redis
redis==5.0.1
celery==5.3.4

# 文档解析
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.2
extract-msg==0.29.1

# HTTP客户端和API
httpx==0.25.2
requests==2.31.0

# 异步处理
asyncio
aioredis==2.0.1

# 工具库
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic-settings==2.0.3
pydantic==2.5.0

# 文件处理
python-magic==0.4.27
Pillow==10.1.0

# 任务调度
APScheduler==3.10.4

# 日志和监控
loguru==0.7.2
sentry-sdk[fastapi]==1.38.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# 开发工具
black==23.11.0
flake8==6.0.0
isort==5.12.0
pre-commit==3.5.0